requirements.txt
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
onnxruntime-silicon==1.16.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.30.2
PyMatting==1.1.13
referencing==0.36.2
rembg==2.0.61
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.24.0
scipy==1.13.1
sympy==1.14.0
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
